Provide option to assume zero initial state in ansatz generation
The ansatz begins with ZXZ decompositions, but the first RZGate
on each qubit has no effect if acting on the initial state.  So,
this adds an option to remove that parameter on each qubit.

It's left as an option since it does not make sure to do in the case
of unitary AQC.

I'm definitely open to discussing the name of the option.

TODO

- [ ] add new argument to docstring