Bump development tests to use Python 3.10 minimum
In recent days, both quimb and Qiskit have dropped support for Python 3.9 on their development branches, and since then the CI cron job has been failing. This fixes CI by bumping to Python 3.10. We may in the future abruptly drop support for Python 3.9, either when we feel the time is right or when there is a reason to bump the required version of either Qiskit or quimb. Until then, people can continue to use Python 3.9, but they are encouraged to upgrade as soon as possible since it is EOL.